agent-a2a/requirements.txt
a2a-sdk>=0.2.0
httpx>=0.27.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
